requirements.txt
accelerate==0.28.0
diffusers==0.27.0
matplotlib==3.7.2
matplotlib==3.7.5
numpy==1.24.1
numpy==1.24.3
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
Pillow==10.3.0
rasterio==1.3.10
torch==2.2.1+cu118
torchsummary==1.5.1
torchview==0.2.6
torchvision==0.17.1+cu118
tqdm==4.66.2
